updated submodules and includeguards are now consistent

+++ engine/core/graphics/format.cpp
@@ -420,9 +420,9 @@ texture_format get_best_format(std::uint16_t type_flags, std::uint32_t search_fl
 	return texture_format::Unknown;
 }
 
-uint32_t get_default_rt_sampler_flags()
+uint64_t get_default_rt_sampler_flags()
 {
-	static std::uint32_t sampler_flags = 0 | BGFX_TEXTURE_RT | BGFX_SAMPLER_U_CLAMP | BGFX_SAMPLER_V_CLAMP;
+	static std::uint64_t sampler_flags = 0 | BGFX_TEXTURE_RT | BGFX_SAMPLER_U_CLAMP | BGFX_SAMPLER_V_CLAMP;
 
 	return sampler_flags;
 }
